requirements.txt
matplotlib>=3.1.2
numpy>=1.17.4
pandas>=0.25.3
scikit-image>=0.16.2
scikit-learn>=0.22
scipy>=1.3.3
SimpleITK>=1.2.4
tensorboard>=2.1.0
tensorboardX>=1.9
torchvision>=1.0.1
PYMIC==0.2.3
